move RAC to open source
